Extract StrokeGameView from the stroke route (Spec 2 Phase 3)
Apply the #545/#546 recipe to stroke — the last format still routing. Move the
stroke game view into a re-hostable component (reads its own tripId + ?game=);
route page becomes a thin wrapper. Pure refactor — no behavior change.

Phase 0 verdict: CLEAN. Stroke's "full-screen overlay" is only its Play/entry
view (identical to match's score sub-screen, which already works in the panel);
its data access, params, hooks, and components are all the modern shared stack —
no pre-competition-era baggage blocking extraction. The one artifact (a
pick-players create screen) is dormant in the panel (always an existing ?game=).
